scripts/train/requirements-train.txt
# Requirements for AI training scripts
# Install with: pip3 install -r requirements-train.txt

# Core ML/AI libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.41.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.3.0

# Text processing
nltk>=3.8.0
spacy>=3.5.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0
python-dotenv>=1.0.0

# Ollama integration
requests>=2.31.0
ollama>=0.1.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0